Update loop count and data type for stress duration
This commit makes two important changes: Firstly, it replaces the hard-coded count of 50 with the numberOfRecords variable in the data generation loops of the heartrate, spo2, and stress examples. This makes the number of iterations flexible and centrally configurable. Secondly, it changes the data type of the HighStressDuration field in the Stress struct and in relevant database structures from int to int64, allowing for more precise tracking of stress duration.

+++ austinapi_db/models.go
@@ -49,7 +49,7 @@ type Spo2 struct {
 type Stress struct {
 	ID                 int64     `json:"id"`
 	Date               time.Time `json:"date"`
-	HighStressDuration int       `json:"high_stress_duration"`
+	HighStressDuration int64     `json:"high_stress_duration"`
 	CreatedTimestamp   time.Time `json:"created_timestamp"`
 	UpdatedTimestamp   time.Time `json:"updated_timestamp"`
 }
